+++ agentrun-20250910/alibabacloud_agentrun20250910/models/_create_memory_collection_input.py
@@ -18,6 +18,7 @@ class CreateMemoryCollectionInput(DaraModel):
         network_configuration: main_models.NetworkConfiguration = None,
         type: str = None,
         vector_store_config: main_models.VectorStoreConfig = None,
+        workspace_id: str = None,
     ):
         self.description = description
         self.embedder_config = embedder_config
@@ -29,6 +30,7 @@ class CreateMemoryCollectionInput(DaraModel):
         self.network_configuration = network_configuration
         self.type = type
         self.vector_store_config = vector_store_config
+        self.workspace_id = workspace_id
 
     def validate(self):
         if self.embedder_config:
@@ -75,6 +77,9 @@ class CreateMemoryCollectionInput(DaraModel):
         if self.vector_store_config is not None:
             result['vectorStoreConfig'] = self.vector_store_config.to_map()
 
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+
         return result
 
     def from_map(self, m: dict = None):
@@ -112,6 +117,9 @@ class CreateMemoryCollectionInput(DaraModel):
         if m.get('vectorStoreConfig') is not None:
             temp_model = main_models.VectorStoreConfig()
             self.vector_store_config = temp_model.from_map(m.get('vectorStoreConfig'))
+
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
 
         return self