+++ agentrun-20250910/alibabacloud_agentrun20250910/models/_memory_collection.py
@@ -20,6 +20,7 @@ class MemoryCollection(DaraModel):
         memory_collection_name: str = None,
         network_configuration: main_models.NetworkConfiguration = None,
         vector_store_config: main_models.VectorStoreConfig = None,
+        workspace_id: str = None,
     ):
         self.created_at = created_at
         self.description = description
@@ -33,6 +34,7 @@ class MemoryCollection(DaraModel):
         self.memory_collection_name = memory_collection_name
         self.network_configuration = network_configuration
         self.vector_store_config = vector_store_config
+        self.workspace_id = workspace_id
 
     def validate(self):
         if self.embedder_config:
@@ -85,6 +87,9 @@ class MemoryCollection(DaraModel):
         if self.vector_store_config is not None:
             result['vectorStoreConfig'] = self.vector_store_config.to_map()
 
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+
         return result
 
     def from_map(self, m: dict = None):
@@ -128,6 +133,9 @@ class MemoryCollection(DaraModel):
         if m.get('vectorStoreConfig') is not None:
             temp_model = main_models.VectorStoreConfig()
             self.vector_store_config = temp_model.from_map(m.get('vectorStoreConfig'))
+
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
 
         return self